Refactor cart context and address lint warnings
