+++ src/components/CartIcon.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { useCart } from '@/context/CartContext';
+import { useCart } from '@/context';
 
 const CartIcon: React.FC = () => {
   const { getTotalItems } = useCart();
