+++ src/context/CartContext.ts
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+import type { CartContextType } from "./cart-types";
+
+export const CartContext = createContext<CartContextType | undefined>(undefined);
