+++ src/context/index.ts
@@ -1,0 +1,3 @@
+export { CartProvider } from "./CartProvider";
+export { useCart } from "./useCart";
+export type { Product, CartContextType } from "./cart-types";
